+++ packages/cli/__tests__/zosfiles/__system__/create/pds/__scripts__/command/command_create_pds_with_primary_secondary.sh
@@ -3,7 +3,7 @@ hlq1=$1
 set -e
 
 echo "================Z/OS FILES CREATE PDS==============="
-zowe zos-files create pds "$1.test.data.set.pds.second" --ps 20 --ss 5 --pa
+zowe zos-files create pds "$1.test.data.set.pds.second" --ps 20 --ss 5 --attributes
 if [ $? -gt 0 ]
 then
     exit $?