+++ packages/zosjobs/src/JobsMessages.ts
@@ -41,5 +41,14 @@ export const ZosJobsMessages: { [key: string]: IMessageDefinition } = {
      */
     missingJcl: {
         message: "No JCL provided"
+    },
+
+    /**
+    * Message indicating that no search option was passed or both were passed
+    * @memberof ZosJobsMessages
+    * @type {IMessageDefinition}
+    */
+    missingSearchOption: {
+        message: "You must specify either the `--search-string` or `--search-regex` option"
     }
 };